+++ python/triton/__init__.py
@@ -1,15 +1,19 @@
 """isort:skip_file"""
 # flake8: noqa: F401
-__version__ = '2.0.0'
+__version__ = "2.0.0"
 
 # TODO: torch needs to be imported first
 # or pybind11 shows `munmap_chunk(): invalid pointer`
 import torch
+
 # submodules
 
 from .utils import *
-from .unroll import *
+from .base import *
 from . import libdevice
+from .core import *
+from .random import *
+from .compiler import *
 
 from . import testing
 from . import ops
